Extracted VmScaler class from BaseCloudConnector module.
